fprintf() wants an int for * format items. (dm)
git-svn-id: svn://mielke.cc/main/brltty@6462 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/tbl2hex.c
@@ -140,7 +140,7 @@ dumpBytes (FILE *stream, const unsigned char *bytes, size_t count) {
   const unsigned char *byte = bytes;
   const unsigned char *end = byte + count;
   int first = 1;
-  size_t digits;
+  int digits;
 
   if (count) {
     char buffer[0X10];
